+++ Cargo.toml
@@ -8,22 +8,22 @@ keywords = ["bevy", "tabletop", "dice"]
 repository = "https://github.com/quentusrex/stackrankdice"
 
 [dependencies]
-bevy_rapier3d = { version = "0.18.0", features = [
+bevy_rapier3d = { version = "0.19.0", features = [
     "simd-stable",
     "debug-render",
 ] }
-bevy_dice = { version = "0.1.3" }
+bevy_dice = { version = "0.2.0" }
 rand = "0.8.5"
 array2d = "0.2.1"
-bevy-inspector-egui = { version = "0.13.0" }
-bevy_mod_outline = "0.2.3"
-bevy_mod_picking = { version = "0.9.0" }
+bevy-inspector-egui = { version = "0.14.0" }
+bevy_mod_outline = { git = "https://github.com/komadori/bevy_mod_outline.git", rev = "5ae478e" }
+bevy_mod_picking = { version = "0.10.0" }
 rand_chacha = "0.3.1"
 clap = { version = "4.0.2", features = ["derive"] }
 
 [dependencies.bevy]
 default-features = false
-version = "0.8.1"
+version = "0.9.0"
 features = [
     "render",
     "animation",
@@ -37,7 +37,9 @@ features = [
 ]
 
 [dependencies.bevy_kira_audio]
-version = "0.12.0"
+#version = "0.13.0"
+git = "https://github.com/NiklasEi/bevy_kira_audio.git"
+rev = "08ec40e"
 default-features = false
 features = ["wav", "ogg"]
