+++ src/events.rs
@@ -8,27 +8,51 @@ use bevy_dice::{DiceRollResult, DiceRollStartEvent};
 use bevy_kira_audio::prelude::*;
 use bevy_mod_picking::{PickingEvent, SelectionEvent};
 
-use crate::game::GameLogEntry;
+use crate::board::{draw_board, StackRankDiceGameBoardElement};
+use crate::game::{GameLogEntry, SelectedRegion};
 use crate::game::{GameState, Region};
 use crate::tiered_prng::PrngMapResource;
-use crate::{
-    draw_board, DiceRollUI, SelectedRegion, StackRankDiceGameBoardElement, StackRankDiceUI,
-};
+use crate::ui::{DiceRollUI, StackRankDiceUI};
 
-pub(crate) struct EventRegionClashStart {
+/// Event that is fired when two regions on a map are entering a clash
+#[allow(dead_code)]
+pub(crate) struct EventPlayerMoveStart {
     region_1: Region,
     region_2: Region,
+    player_1: usize,
+    player_2: usize,
 }
 
-pub(crate) struct EventRegionClashEnd {
-    region1: Region,
-    region2: Region,
-    dice_1_sum: usize,
-    dice_2_sum: usize,
+/// Event that is fired when a clash between two regions on a map is resolved
+/// and the winner is determined
+#[allow(dead_code)]
+pub(crate) struct EventPlayerMoveEnd {
+    player_1: usize,
+    player_2: usize,
+    region_1: Region,
+    region_2: Region,
+    region_1_dice_result: Vec<usize>,
+    region_2_dice_result: Vec<usize>,
 }
 
+/// Event that is fired when a played has won a game
 pub(crate) struct EventGameOver {
+    // An index of a winner
     winner: usize,
+}
+
+/// Event that is fired when a turn of a player is started
+#[allow(dead_code)]
+pub(crate) struct EventTurnStart {
+    // An index of a player
+    player: usize,
+}
+
+/// Event that is fired when a turn of a player is started
+#[allow(dead_code)]
+pub(crate) struct EventTurnEnd {
+    // An index of a player
+    player: usize,
 }
 
 pub(crate) fn filter_just_selected_event(
@@ -48,7 +72,7 @@ pub(crate) fn event_region_selected(
     picking_events: EventReader<PickingEvent>,
     regions: Query<(Entity, &Region)>,
     game_state: Res<GameState>,
-    mut event_writer: EventWriter<EventRegionClashStart>,
+    mut event_writer: EventWriter<EventPlayerMoveStart>,
 ) {
     let selected_entity = filter_just_selected_event(picking_events);
 
@@ -64,7 +88,12 @@ pub(crate) fn event_region_selected(
             let region_2 = region.clone();
             if region_1.is_opponent(&region_2) {
                 // Attack a neighbour
-                let event = EventRegionClashStart { region_1, region_2 };
+                let event = EventPlayerMoveStart {
+                    player_1: region_1.owner,
+                    player_2: region_2.owner,
+                    region_1,
+                    region_2,
+                };
                 event_writer.send(event);
             }
         }
@@ -80,9 +109,9 @@ pub(crate) struct DiceRollTimer {
     timer: Timer,
 }
 
-pub(crate) fn event_region_clash(
+pub(crate) fn event_player_move_start(
     mut commands: Commands,
-    mut region_clash_event_reader: EventReader<EventRegionClashStart>,
+    mut region_clash_event_reader: EventReader<EventPlayerMoveStart>,
     mut dice_roll_started_writer: EventWriter<DiceRollStartEvent>,
     mut dice_roll_view_query: Query<(Entity, &mut Visibility, &DiceRollUI)>,
     mut game_state: ResMut<GameState>,
@@ -107,15 +136,15 @@ pub(crate) fn event_region_clash(
             turn_of_player,
             region_1: event.region_1.clone(),
             region_2: event.region_2.clone(),
-            dice_1_sum: 0,
-            dice_2_sum: 0,
+            region_1_dice_result: Vec::new(),
+            region_2_dice_result: Vec::new(),
             turn_counter,
         });
 
         dice_roll_started_writer.send(dice_roll_started);
 
-        commands.spawn().insert(DiceRollTimer {
-            timer: Timer::new(Duration::from_secs(3), false),
+        commands.spawn(()).insert(DiceRollTimer {
+            timer: Timer::new(Duration::from_secs(3), TimerMode::Once),
         });
     }
 }
@@ -131,8 +160,8 @@ pub(crate) fn event_dice_roll_result(
 
         audio.play(asset_server.load("sounds/throw.wav"));
 
-        last_log_entry.dice_1_sum = event.values[0].iter().sum();
-        last_log_entry.dice_2_sum = event.values[1].iter().sum();
+        last_log_entry.region_1_dice_result = event.values[0].clone();
+        last_log_entry.region_2_dice_result = event.values[1].clone();
     }
 }
 
@@ -141,7 +170,7 @@ pub(crate) fn event_dice_rolls_complete(
     mut dice_roll_timer_query: Query<(Entity, &mut DiceRollTimer)>,
     mut dice_roll_ui_query: Query<(Entity, &mut Visibility, &mut DiceRollUI)>,
     time: Res<Time>,
-    mut region_clash_end_event_writer: EventWriter<EventRegionClashEnd>,
+    mut region_clash_end_event_writer: EventWriter<EventPlayerMoveEnd>,
     mut game_state: ResMut<GameState>,
 ) {
     for (entity, mut fuse_timer) in dice_roll_timer_query.iter_mut() {
@@ -155,19 +184,21 @@ pub(crate) fn event_dice_rolls_complete(
 
             let last_log_entry = game_state.game_log.last_mut().unwrap();
 
-            region_clash_end_event_writer.send(EventRegionClashEnd {
-                region1: last_log_entry.region_1.clone(),
-                region2: last_log_entry.region_2.clone(),
-                dice_1_sum: last_log_entry.dice_1_sum,
-                dice_2_sum: last_log_entry.dice_2_sum,
+            region_clash_end_event_writer.send(EventPlayerMoveEnd {
+                player_1: last_log_entry.region_1.owner,
+                player_2: last_log_entry.region_2.owner,
+                region_1: last_log_entry.region_1.clone(),
+                region_2: last_log_entry.region_2.clone(),
+                region_1_dice_result: last_log_entry.region_1_dice_result.clone(),
+                region_2_dice_result: last_log_entry.region_2_dice_result.clone(),
             })
         }
     }
 }
 
 #[allow(clippy::too_many_arguments)]
-pub(crate) fn event_region_clash_end(
-    mut region_clash_end_event_reader: EventReader<EventRegionClashEnd>,
+pub(crate) fn event_player_move_end(
+    mut region_clash_end_event_reader: EventReader<EventPlayerMoveEnd>,
     mut game_state: ResMut<GameState>,
     mut game_elements_query: Query<(Entity, &StackRankDiceGameBoardElement)>,
     asset_server: Res<AssetServer>,
@@ -178,30 +209,35 @@ pub(crate) fn event_region_clash_end(
     mut selected_region: ResMut<SelectedRegion>,
     audio: Res<bevy_kira_audio::prelude::Audio>,
     mut event_game_over_writer: EventWriter<EventGameOver>,
+    mut event_turn_end_writer: EventWriter<EventTurnEnd>,
+    mut event_turn_start_writer: EventWriter<EventTurnStart>,
 ) {
     let mut rng = rand::thread_rng();
     let mut redraw_board = false;
 
     for e in region_clash_end_event_reader.iter() {
-        if e.dice_1_sum > e.dice_2_sum {
+        let result_1: usize = e.region_1_dice_result.iter().sum();
+        let result_2: usize = e.region_2_dice_result.iter().sum();
+
+        if result_1 > result_2 {
             // win a region
-            game_state.board.regions[e.region2.id].owner = e.region1.owner;
-            if e.region1.num_dice > 1 {
-                game_state.board.regions[e.region2.id].num_dice =
-                    rng.gen_range(1..e.region1.num_dice);
-                game_state.board.regions[e.region1.id].num_dice -=
-                    game_state.board.regions[e.region2.id].num_dice - 1;
+            game_state.board.regions[e.region_2.id].owner = e.region_1.owner;
+            if e.region_1.num_dice > 1 {
+                game_state.board.regions[e.region_2.id].num_dice =
+                    rng.gen_range(1..e.region_1.num_dice);
+                game_state.board.regions[e.region_1.id].num_dice -=
+                    game_state.board.regions[e.region_2.id].num_dice - 1;
             }
 
             audio.play(asset_server.load("sounds/win.wav"));
         } else {
             // lose a region
-            game_state.board.regions[e.region1.id].owner = e.region2.owner;
-            if e.region2.num_dice > 1 {
-                game_state.board.regions[e.region1.id].num_dice =
-                    rng.gen_range(1..e.region2.num_dice);
-                game_state.board.regions[e.region2.id].num_dice -=
-                    game_state.board.regions[e.region1.id].num_dice - 1;
+            game_state.board.regions[e.region_1.id].owner = e.region_2.owner;
+            if e.region_2.num_dice > 1 {
+                game_state.board.regions[e.region_1.id].num_dice =
+                    rng.gen_range(1..e.region_2.num_dice);
+                game_state.board.regions[e.region_2.id].num_dice -=
+                    game_state.board.regions[e.region_1.id].num_dice - 1;
             }
 
             audio.play(asset_server.load("sounds/loss.wav"));
@@ -248,10 +284,19 @@ pub(crate) fn event_region_clash_end(
         .count();
 
     if number_of_unblocked_regions == 0 {
+        event_turn_end_writer.send(EventTurnEnd {
+            player: game_state.turn_of_player,
+        });
+
         game_state.turn_of_player += 1;
         if game_state.turn_of_player >= game_state.number_of_players {
             game_state.turn_of_player = 0;
         }
+
+        event_turn_start_writer.send(EventTurnStart {
+            player: game_state.turn_of_player,
+        });
+
         game_state.turn_counter += 1;
     }
 
@@ -300,7 +345,7 @@ pub(crate) fn event_game_over(
         }
 
         commands
-            .spawn_bundle(
+            .spawn(
                 TextBundle::from_section(
                     format!("Player {} wins!", e.winner + 1),
                     TextStyle {
@@ -322,6 +367,6 @@ pub(crate) fn event_game_over(
             )
             .insert(StackRankDiceUI);
 
-        // audio.play(asset_server.load("sounds/game_over.wav"));
+        // _audio.play(asset_server.load("sounds/game_over.wav"));
     }
 }
